+++ lib/src/analyzers/lint_analyzer/reporters/reporters_list/code_climate/lint_code_climate_reporter.dart
@@ -9,12 +9,14 @@ import '../../../models/issue.dart';
 import '../../../models/lint_file_report.dart';
 import '../../../models/report.dart';
 import '../../../models/severity.dart';
+import '../../../models/summary_lint_report_record.dart';
 import 'models/code_climate_issue.dart';
 import 'models/code_climate_issue_category.dart';
 import 'models/code_climate_issue_location.dart';
 import 'models/code_climate_issue_severity.dart';
 
-class LintCodeClimateReporter extends CodeClimateReporter<LintFileReport> {
+class LintCodeClimateReporter
+    extends CodeClimateReporter<LintFileReport, SummaryLintReportRecord> {
   LintCodeClimateReporter(IOSink output, {bool gitlabCompatible = false})
       : super(
           output,
@@ -22,7 +24,10 @@ class LintCodeClimateReporter extends CodeClimateReporter<LintFileReport> {
         );
 
   @override
-  Future<void> report(Iterable<LintFileReport> records) async {
+  Future<void> report(
+    Iterable<LintFileReport> records, {
+    Iterable<SummaryLintReportRecord> summary = const [],
+  }) async {
     if (records.isEmpty) {
       return;
     }