+++ lib/src/analyzers/unused_files_analyzer/reporters/reporters_list/json/unused_files_json_reporter.dart
@@ -7,11 +7,15 @@ import '../../../../../reporters/models/json_reporter.dart';
 import '../../../models/unused_files_file_report.dart';
 
 @immutable
-class UnusedFilesJsonReporter extends JsonReporter<UnusedFilesFileReport> {
+class UnusedFilesJsonReporter
+    extends JsonReporter<UnusedFilesFileReport, void> {
   const UnusedFilesJsonReporter(IOSink output) : super(output, 2);
 
   @override
-  Future<void> report(Iterable<UnusedFilesFileReport> records) async {
+  Future<void> report(
+    Iterable<UnusedFilesFileReport> records, {
+    Iterable<void> summary = const [],
+  }) async {
     if (records.isEmpty) {
       return;
     }