+++ lib/src/reporters/models/console_reporter.dart
@@ -6,7 +6,7 @@ import 'file_report.dart';
 import 'reporter.dart';
 
 /// Plain terminal reporter
-abstract class ConsoleReporter<T extends FileReport> extends Reporter<T> {
+abstract class ConsoleReporter<T extends FileReport, S> extends Reporter<T, S> {
   static const String id = 'console';
   static const String verboseId = 'console-verbose';