+++ lib/src/reporters/models/github_reporter.dart
@@ -7,7 +7,7 @@ import 'file_report.dart';
 import 'reporter.dart';
 
 /// Creates report about issues in pull request based on GitHub Actions Workflow commands.
-abstract class GitHubReporter<T extends FileReport> extends Reporter<T> {
+abstract class GitHubReporter<T extends FileReport, S> extends Reporter<T, S> {
   static const String id = 'github';
 
   static final commands = GitHubWorkflowCommands();