+++ lib/src/reporters/models/json_reporter.dart
@@ -6,7 +6,7 @@ import 'file_report.dart';
 import 'reporter.dart';
 
 /// Machine-readable report in JSON format
-abstract class JsonReporter<T extends FileReport> extends Reporter<T> {
+abstract class JsonReporter<T extends FileReport, S> extends Reporter<T, S> {
   static const String id = 'json';
 
   @protected